@@ -1,18 +1,20 @@
 module github.com/Depado/goploader
 
-go 1.18
+go 1.23.0
+
+toolchain go1.24.0
 
 require (
+	dario.cat/mergo v1.0.1
 	github.com/asdine/storm v2.1.2+incompatible
 	github.com/atotto/clipboard v0.1.4
-	github.com/dchest/uniuri v0.0.0-20200228104902-7aecb25e1fe5
-	github.com/fatih/color v1.13.0
-	github.com/gin-gonic/gin v1.9.1
-	github.com/imdario/mergo v0.3.12
+	github.com/dchest/uniuri v1.2.0
+	github.com/fatih/color v1.18.0
+	github.com/gin-gonic/gin v1.10.0
 	github.com/mitchellh/go-homedir v1.1.0
 	github.com/ogier/pflag v0.0.1
-	github.com/prometheus/client_golang v1.12.1
-	go.etcd.io/bbolt v1.3.6
+	github.com/prometheus/client_golang v1.21.0
+	go.etcd.io/bbolt v1.4.0
 	gopkg.in/cheggaaa/pb.v1 v1.0.28
 	gopkg.in/yaml.v2 v2.4.0
 )
