@@ -1,7 +1,7 @@
 package conf
 
 import (
-	"io/ioutil"
+	"fmt"
 	"log"
 	"os"
 
@@ -34,22 +34,24 @@ func Load() error {
 
 	if _, err = os.Stat(cdir); os.IsNotExist(err) {
 		log.Printf("Creating %v directory.\n", cdir)
-		os.Mkdir(cdir, 0700)
+		if err := os.Mkdir(cdir, 0700); err != nil {
+			return fmt.Errorf("unable to create %v directory: %w", cdir, err)
+		}
 	} else if err != nil {
 		return err
 	}
 	if _, err = os.Stat(cf); os.IsNotExist(err) {
 		log.Printf("Configuration file %v not found. Writing default configuration.\n", cf)
-		C.Service = "https://gpldr.in/"
+		C.Service = "http://127.0.0.1:8080"
 		if conf, err = yaml.Marshal(C); err != nil {
 			return err
 		}
-		return ioutil.WriteFile(cf, conf, 0644)
+		return os.WriteFile(cf, conf, 0644)
 	} else if err != nil {
 		return err
 	}
 
-	if conf, err = ioutil.ReadFile(cf); err != nil {
+	if conf, err = os.ReadFile(cf); err != nil {
 		return err
 	}
 	return yaml.Unmarshal(conf, &C)
